requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
djangorestframework==3.12.2
Faker==4.14.2
idna==2.10
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
